requirements.txt
faiss-cpu
transformers
torch
bs4
requests
numpy